pyproject.toml
[project]
name = "gamelib"
version = "0.1.1"
description = ""
authors = [
    {name = "welc0186",email = "49768046+welc0186@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"

[tool.poetry]
packages = [{include = "gamelib"}]

[tool.poetry.dependencies]
python = ">=3.11"
esper = ">=3.4,<4.0"
pygame-ce = "^2.5.6"
pytmx = "^3.32"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
ruff = "^0.14.14"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
